### 2025-02-14 #7778 @danielskatz
👋 @jan-janssen - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

In addition, JOSS papers are intended to be a maximum of about 1000 words, and your paper is close to 3000, so a lot will need to be cut. You can use the command `@editorialbot check repository` to see what JOSS thinks the word could is after you make changes in your paper.  On the other hand, if you think this amount of text is needed, you would need to withdraw this submission and find another venue for the work.
