+++ docs/paper/paper.bib
@@ -89,9 +89,10 @@
     author = {Matthew Rocklin},
     title = {Dask: Parallel Computation with Blocked algorithms and Task Scheduling},
     booktitle = {Proceedings of the 14th Python in Science Conference},
-    pages = {130 - 136},
+    pages = {126 - 132},
     year = {2015},
-    editor = {Kathryn Huff and James Bergstra}
+    editor = {Kathryn Huff and James Bergstra},
+    doi = {10.25080/Majora-7b98e3ed-013}
 }
 
 @inproceedings{parsl, 
