### 2025-02-15 #7778 @jan-janssen
@danielskatz based on the JOSS database, I would suggest the following group of five reviewers:
* Rohit Goswami - Github: HaoZeke
* Gaurav Sablok - Github: sablokgaurav
* Baodi Shan - Github: lwshanbd
* Sven Berger - Github: svchb
* Marcin Rogowski  - Github: mrogowski
