### 2025-02-15 ea691a4
Refactor interactive argument handling
