+++ executorlib/interactive/shared.py
@@ -152,6 +152,7 @@ class InteractiveExecutor(ExecutorBroker):
         executor_kwargs["future_queue"] = self._future_queue
         executor_kwargs["spawner"] = spawner
         executor_kwargs["queue_join_on_shutdown"] = False
+        self._process_kwargs = executor_kwargs
         self._set_process(
             process=[
                 Thread(
@@ -209,6 +210,7 @@ class InteractiveStepExecutor(ExecutorBase):
         executor_kwargs["spawner"] = spawner
         executor_kwargs["max_cores"] = max_cores
         executor_kwargs["max_workers"] = max_workers
+        self._process_kwargs = executor_kwargs
         self._set_process(
             Thread(
                 target=execute_separate_tasks,
