### 2025-02-15 1c77bbd
refactor dict
