### 2025-02-15 3befb3f
Attach process kwargs to Executor (#586)
* Attach process kwargs to Executor

* refactor dict

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fixes

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
