### 2025-02-15 e248ffa
Refactor interactive executor classes
