+++ tests/test_shell_interactive.py
@@ -6,7 +6,7 @@ import unittest
 
 from executorlib import SingleNodeExecutor
 from executorlib.standalone.serialize import cloudpickle_register
-from executorlib.interactive.shared import execute_parallel_tasks
+from executorlib.interactive.shared import execute_tasks
 from executorlib.standalone.interactive.spawner import MpiExecSpawner
 
 
@@ -88,7 +88,7 @@ class ShellInteractiveExecutorTest(unittest.TestCase):
         cloudpickle_register(ind=1)
         self.assertFalse(future_lines.done())
         self.assertFalse(future_pattern.done())
-        execute_parallel_tasks(
+        execute_tasks(
             future_queue=test_queue,
             cores=1,
             openmpi_oversubscribe=False,

[…]
