+++ tests/test_pysqa_subprocess.py
@@ -1,5 +1,5 @@
 import unittest
-from executorlib.interactive.slurm import generate_slurm_command
+from executorlib.interactive.slurmspawner import generate_slurm_command
 
 try:
     from executorlib.cache.queue_spawner import _pysqa_execute_command
