+++ executorlib/interactive/fluxspawner.py
@@ -1,9 +1,6 @@
 import os
 from typing import Optional
 
-import fluxspawner
-import fluxspawner.job
-
 from executorlib.standalone.interactive.spawner import BaseSpawner
 
 
