+++ executorlib/interfaces/flux.py
@@ -15,7 +15,10 @@ from executorlib.standalone.inputcheck import (
 )
 
 with contextlib.suppress(ImportError):
-    from executorlib.interactive.fluxspawner import FluxPythonSpawner, validate_max_workers
+    from executorlib.interactive.fluxspawner import (
+        FluxPythonSpawner,
+        validate_max_workers,
+    )
 
 
 class FluxJobExecutor:
