+++ executorlib/interactive/fluxspawner.py
@@ -1,6 +1,9 @@
 import os
 from typing import Optional
 
+import flux
+import flux.job
+
 from executorlib.standalone.interactive.spawner import BaseSpawner
 
 
