+++ tests/test_executor_backend_flux.py
@@ -8,7 +8,7 @@ from executorlib import FluxJobExecutor
 
 try:
     import flux.job
-    from executorlib.interactive.flux import FluxPythonSpawner
+    from executorlib.interactive.fluxspawner import FluxPythonSpawner
 
     skip_flux_test = "FLUX_URI" not in os.environ
     pmi = os.environ.get("EXECUTORLIB_PMIX", None)

[…]
