### 2025-02-15 31f5f13
Generate UML diagram
