+++ .github/workflows/pipeline.yml
@@ -405,6 +405,8 @@ jobs:
       timeout-minutes: 10
       run: |
         pyreverse -o png -p ./${{ github.event.repository.name }} ./${{ github.event.repository.name }}
+        ls *.png
+        ls ./${{ github.event.repository.name }}/*.png
     - uses: actions/upload-artifact@v4
       with:
         name: ${{ github.event.repository.name }}.png
