### 2025-02-15 00366ad
Generate UML diagram (#588)
* Generate UML diagram

* comment out needs for testing

* fix path

* look for files

* Update pipeline.yml

* Update pipeline.yml
