### 2025-02-15 #7778 @danielskatz
👋 @sciencegenome, @svchb, @mrogowski - would one or more of you be willing to review this submission for JOSS?
