### 2025-02-15 #7778 @danielskatz
👋 @yadudoc & @benclifford - would one or both of you be willing to review this submission for JOSS?
