### 2025-02-15 #7778 @lwshanbd
> 👋  @sablokgaurav & @lwshanbd, would one or both of you be willing to review this submission for JOSS?

Hi, i'd like to review it. Thanks
