### 2025-02-15 1d36d17
[feature] resize block allocation executor
