+++ executorlib/interactive/blockallocation.py
@@ -80,9 +80,7 @@ class BlockAllocationExecutor(ExecutorBase):
                 self._future_queue.queue.insert(0, {"shutdown": True, "wait": True})
             while len(self._process) > max_workers:
                 self._process = [
-                    process
-                    for process in self._process
-                    if process.is_alive()
+                    process for process in self._process if process.is_alive()
                 ]
         elif self._max_workers < max_workers:
             new_process_lst = [
