+++ executorlib/interactive/blockallocation.py
@@ -1,6 +1,7 @@
 from concurrent.futures import Future
 from threading import Thread
 from typing import Callable, Optional
+import queue
 
 from executorlib.base.executor import ExecutorBase, cancel_items_in_queue
 from executorlib.interactive.shared import execute_tasks
@@ -75,27 +76,26 @@ class BlockAllocationExecutor(ExecutorBase):
 
     @max_workers.setter
     def max_workers(self, max_workers: int):
-        if self._max_workers > max_workers:
-            for _ in range(self._max_workers - max_workers):
-                self._future_queue.queue.insert(0, {"shutdown": True, "wait": True})
-            if isinstance(self._process, list):
+        if isinstance(self._future_queue, queue.Queue) and isinstance(self._process, list):
+            if self._max_workers > max_workers:
+                for _ in range(self._max_workers - max_workers):
+                    self._future_queue.queue.insert(0, {"shutdown": True, "wait": True})
                 while len(self._process) > max_workers:
                     self._process = [
                         process for process in self._process if process.is_alive()
                     ]
-        elif self._max_workers < max_workers:
-            new_process_lst = [
-                Thread(
-                    target=execute_tasks,
-                    kwargs=self._process_kwargs,
-                )
-                for _ in range(max_workers - self._max_workers)
-            ]
-            for process_instance in new_process_lst:
-                process_instance.start()
-            if isinstance(self._process, list):
+            elif self._max_workers < max_workers:
+                new_process_lst = [
+                    Thread(
+                        target=execute_tasks,
+                        kwargs=self._process_kwargs,
+                    )
+                    for _ in range(max_workers - self._max_workers)
+                ]
+                for process_instance in new_process_lst:
+                    process_instance.start()
                 self._process += new_process_lst
-        self._max_workers = max_workers
+            self._max_workers = max_workers
 
     def submit(  # type: ignore
         self, fn: Callable, *args, resource_dict: Optional[dict] = None, **kwargs
