+++ tests/test_local_executor_resize.py
@@ -1,5 +1,6 @@
 import unittest
 from executorlib import SingleNodeExecutor
+from executorlib.standalone.serialize import cloudpickle_register
 
 
 def sleep_funct(sec):
@@ -10,6 +11,7 @@ def sleep_funct(sec):
 
 class TestResizing(unittest.TestCase):
     def test_without_dependencies_decrease(self):
+        cloudpickle_register(ind=1)
         with SingleNodeExecutor(max_workers=2, block_allocation=True, disable_dependencies=True) as exe:
             future_lst = [exe.submit(sleep_funct, 1) for _ in range(4)]
             self.assertEqual([f.done() for f in future_lst], [False, False, False, False])
@@ -23,6 +25,7 @@ class TestResizing(unittest.TestCase):
             self.assertEqual([f.done() for f in future_lst], [True, True, True, True])
 
     def test_without_dependencies_increase(self):
+        cloudpickle_register(ind=1)
         with SingleNodeExecutor(max_workers=1, block_allocation=True, disable_dependencies=True) as exe:
             future_lst = [exe.submit(sleep_funct, 0.1) for _ in range(4)]
             self.assertEqual([f.done() for f in future_lst], [False, False, False, False])
