### 2025-02-15 04174e7
Update tests/test_dependencies_executor.py
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
