+++ executorlib/interactive/dependency.py
@@ -73,12 +73,12 @@ class DependencyExecutor(ExecutorBase):
         Returns:
             Optional[dict]: Information about the executor.
         """
-        f: Future = Future()
         if isinstance(self._future_queue, queue.Queue):
+            f: Future = Future()
             self._future_queue.queue.insert(
                 0, {"internal": True, "task": "info", "future": f}
             )
-        return f.result()
+            return f.result()
 
     def submit(  # type: ignore
         self,
