+++ tests/test_local_executor_resize.py
@@ -34,8 +34,39 @@ class TestResizing(unittest.TestCase):
             future_lst[0].result()
             exe.max_workers = 2
             self.assertEqual(exe.max_workers, 2)
-            self.assertEqual(len(exe), 2)
+            self.assertTrue(len(exe) >= 1)
             self.assertEqual(len(exe._process), 2)
             self.assertEqual([f.done() for f in future_lst], [True, False, False, False])
             self.assertEqual([f.result() for f in future_lst], [0.1, 0.1, 0.1, 0.1])
             self.assertEqual([f.done() for f in future_lst], [True, True, True, True])
+
+    def test_with_dependencies_decrease(self):
+        cloudpickle_register(ind=1)
+        with SingleNodeExecutor(max_workers=2, block_allocation=True, disable_dependencies=False) as exe:
+            future_lst = [exe.submit(sleep_funct, 1) for _ in range(4)]
+            self.assertEqual([f.done() for f in future_lst], [False, False, False, False])
+            self.assertEqual(len(exe), 4)
+            sleep_funct(sec=0.5)
+            exe.max_workers = 1
+            self.assertEqual([f.done() for f in future_lst], [True, True, False, False])
+            self.assertEqual([f.result() for f in future_lst], [1, 1, 1, 1])
+            self.assertEqual([f.done() for f in future_lst], [True, True, True, True])
+
+    def test_with_dependencies_increase(self):
+        cloudpickle_register(ind=1)
+        with SingleNodeExecutor(max_workers=1, block_allocation=True, disable_dependencies=False) as exe:
+            future_lst = [exe.submit(sleep_funct, 0.1) for _ in range(4)]
+            self.assertEqual([f.done() for f in future_lst], [False, False, False, False])
+            self.assertEqual(len(exe), 4)
+            self.assertEqual(exe.max_workers, 1)
+            future_lst[0].result()
+            exe.max_workers = 2
+            self.assertEqual(exe.max_workers, 2)
+            self.assertEqual([f.done() for f in future_lst], [True, False, False, False])
+            self.assertEqual([f.result() for f in future_lst], [0.1, 0.1, 0.1, 0.1])
+            self.assertEqual([f.done() for f in future_lst], [True, True, True, True])
+
+    def test_no_block_allocation(self):
+        with self.assertRaises(NotImplementedError):
+            with SingleNodeExecutor(block_allocation=False) as exe:
+                exe.max_workers = 2
