### 2025-02-15 6022dec
fixes and increased test coverage
