### 2025-02-15 8860291
[feature] resize block allocation executor (#589)
* [feature] resize block allocation executor

* extend test

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fixes

* fix type check

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fixes

* test property

* test with dependencies

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fixes and increased test coverage

* fix base property

---------

Co-authored-by: pre-commit
[…]
