+++ executorlib/base/executor.py
@@ -33,6 +33,14 @@ class ExecutorBase(FutureExecutor):
         self._process: Optional[Union[Thread, list[Thread]]] = None
 
     @property
+    def max_workers(self) -> Optional[int]:
+        return self._process_kwargs.get("max_workers")
+
+    @max_workers.setter
+    def max_workers(self, max_workers: int):
+        raise NotImplementedError("The max_workers setter is not implemented.")
+
+    @property
     def info(self) -> Optional[dict]:
         """
         Get the information about the executor.
