### 2025-02-15 #7778 @sciencegenome
> 👋 [@sciencegenome](https://github.com/sciencegenome), [@svchb](https://github.com/svchb), [@mrogowski](https://github.com/mrogowski) - would one or more of you be willing to review this submission for JOSS?

@danielskatz i will review the paper. Sorry for the delay in reply. I am accessing github only in the evening.
