### 2025-02-15 #7778 @sciencegenome
@danielskatz let me know as how to review the paper and submit the comments. Any link that i should be reading and i shall be able to do this by Monday.
