### 2025-02-15 #7778 @sciencegenome
@danielskatz @jan-janssen let me know as how to review the paper and submit the comments. Any link that i should be reading and i shall be able to do this by Monday. I did look at the code just now and this is a very good approach. However, how this can be different from the python greenlet  https://github.com/python-greenlet/greenlet/ .

It is basically still lockig the shared thread and does it follow the principle of the work stealing as if the thread is not available then automatic shifting of the concurrency to the next thread.

Good evening and regards,
Gaurav
