### 2025-02-15 #7778 @danielskatz
@sciencegenome - I'll add you in the system, but I want to get one more reviewer before we start the review.  Regarding the review itself, once it starts, it will follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
