### 2025-02-15 #7782 @editorialbot
[REVIEW]: Executorlib – Up-scaling Python workflows for hierarchical heterogenous high-performance computing
**Submitting author:** @jan-janssen (<a href="http://orcid.org/0000-0001-9948-7119">Jan Janssen</a>)
**Repository:** https://github.com/pyiron/executorlib
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.0.0
