### 2025-02-15 #7782 @svchb
@jan-janssen You list 5 authors of which 4 have not participated actively in development, which for JOSS Papers is in my experience unusual.
