### 2025-02-18 #7782 @jan-janssen
> [@jan-janssen](https://github.com/jan-janssen) You list 5 authors of which 4 have not participated actively in development, which for JOSS Papers is in my experience unusual.

Hi @svchb ,

Thank you for the feedback. While I did the development work, this was in active collaboration with the Co-authors:
* Michael did the testing of the different prototypes and gave valuable feedback to improve the stability of executorlib.
* Ping helped with her feedback to extend the concept from a specialized tool for Exascale HPC to HPC in general, which allows us to target a much larger user base compared to the one we initially intended.
* Joerg helped with his feedback to improve the user interface, highlighting the need for intuitive interfaces in particular for technically less experienced users.
* Danny with his expertise in Exascale HPC was a key driver in optimizing the performance in executorlib.

So given executorlib would not be at the stage where it is now without the whole team, I would like to include them all as authors on the publication. Or are the authors of JOSS papers limited to only the authors who contributed code?

Best,

Jan
