### 2025-02-19 c197736
Add test for priority in flux
