+++ .ci_support/environment-docs.yml
@@ -12,5 +12,5 @@ dependencies:
 - pyzmq =26.2.1
 - flux-core
 - jupyter-book =1.0.0
-- h5py =3.12.1
+- h5py =3.13.0
 - python =3.12
