+++ pyproject.toml
@@ -36,7 +36,7 @@ Documentation = "https://executorlib.readthedocs.io"
 Repository = "https://github.com/pyiron/executorlib"
 
 [project.optional-dependencies]
-cache = ["h5py==3.12.1"]
+cache = ["h5py==3.13.0"]
 graph = [
     "pygraphviz==1.14",
     "networkx==3.4.2",
