### 2025-02-19 4f9ef5f
h5py - 3.13.0 (#593)
* Update pyproject.toml

* Update environment-openmpi.yml

* Update environment-mpich.yml

* Update environment-win.yml

* Update environment-docs.yml

* Update pyproject.toml
