+++ .ci_support/environment-win.yml
@@ -7,7 +7,7 @@ dependencies:
 - cloudpickle =3.1.1
 - mpi4py =4.0.1
 - pyzmq =26.2.1
-- h5py =3.12.1
+- h5py =3.13.0
 - networkx =3.4.2
 - pygraphviz =1.14
 - ipython =8.32.0
