### 2025-02-19 baf63ad
Bump mpi4py from 4.0.1 to 4.0.3
Bumps [mpi4py](https://github.com/mpi4py/mpi4py) from 4.0.1 to 4.0.3.
- [Release notes](https://github.com/mpi4py/mpi4py/releases)
- [Changelog](https://github.com/mpi4py/mpi4py/blob/master/CHANGES.rst)
- [Commits](https://github.com/mpi4py/mpi4py/compare/4.0.1...4.0.3)

---
updated-dependencies:
- dependency-name: mpi4py
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
