+++ binder/environment.yml
@@ -5,7 +5,7 @@ dependencies:
 - numpy
 - openmpi
 - cloudpickle =3.1.1
-- mpi4py =4.0.1
+- mpi4py =4.0.3
 - pyzmq =26.2.1
 - flux-core =0.59.0
 - flux-pmix =0.5.0
