### 2025-02-19 pyiron/executorlib#594 @svchb
Joss Review Software Paper
# Statement of Need

> ~Previously, the Python programming language was primarily used in scientific HPC workloads to couple
> performance-critical scientific software packages written in different programming languages in
> order to solve complex tasks.~

This is too much of a generalization and not really important

> like the SLURM

such as SLURM

> While these tools can be powerful, they introduce new constructs that are not familiar to most Python developers, introducing complexity
> and a barrier to entry.

While these tools can be powerful, they introduce new constructs that unfamiliar to most Python developers, adding complexity and creating a barrier to entry.

Please name the software packages that you compare to **executorlib**, i.e. Aiida, myqueue... just the references is hard to read. Also try to formulate the differences a bit more with 1-2 additional sentences.

# Features and Implementation

>  the powerful novel hierarchical

the powerful and novel hierarchical

**Please extend the functional description (~100 words).**
