### 2025-02-19 #7782 @svchb
Paper issue thread: https://github.com/pyiron/executorlib/issues/594
