### 2025-02-28 f8c8a85
Bump ipython from 8.32.0 to 8.33.0
Bumps [ipython](https://github.com/ipython/ipython) from 8.32.0 to 8.33.0.
- [Release notes](https://github.com/ipython/ipython/releases)
- [Commits](https://github.com/ipython/ipython/compare/8.32.0...8.33.0)

---
updated-dependencies:
- dependency-name: ipython
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
