+++ .ci_support/environment-win.yml
@@ -10,5 +10,5 @@ dependencies:
 - h5py =3.13.0
 - networkx =3.4.2
 - pygraphviz =1.14
-- ipython =8.32.0
+- ipython =8.33.0
 - setuptools =75.1.0
