+++ pyproject.toml
@@ -46,13 +46,13 @@ graphnotebook = [
     "networkx==3.4.2",
     "ipython==8.33.0",
 ]
-mpi = ["mpi4py==4.0.1"]
+mpi = ["mpi4py==4.0.3"]
 cluster = [
     "pysqa==0.2.3",
     "h5py==3.13.0",
 ]
 all = [
-    "mpi4py==4.0.1",
+    "mpi4py==4.0.3",
     "pysqa==0.2.3",
     "h5py==3.13.0",
     "pygraphviz==1.14",
