### 2025-03-03 8fb04ee
Bump ipython from 8.33.0 to 9.0.1
Bumps [ipython](https://github.com/ipython/ipython) from 8.33.0 to 9.0.1.
- [Release notes](https://github.com/ipython/ipython/releases)
- [Commits](https://github.com/ipython/ipython/compare/8.33.0...9.0.1)

---
updated-dependencies:
- dependency-name: ipython
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
