+++ binder/environment.yml
@@ -15,4 +15,4 @@ dependencies:
 - networkx =3.4.2
 - pygraphviz =1.14
 - pysqa =0.2.3
-- ipython =8.33.0
+- ipython =9.0.1
