+++ .github/workflows/pipeline.yml
@@ -259,9 +259,6 @@ jobs:
         - operating-system: ubuntu-latest
           python-version: '3.11'
 
-        - operating-system: ubuntu-latest
-          python-version: '3.10'
-
     steps:
     - uses: actions/checkout@v4
     - name: Conda config
@@ -299,9 +296,6 @@ jobs:
 
         - operating-system: ubuntu-latest
           python-version: '3.11'
-
-        - operating-system: ubuntu-latest
-          python-version: '3.10'
 
     steps:
     - uses: actions/checkout@v4
