+++ .ci_support/environment-openmpi.yml
@@ -5,7 +5,7 @@ dependencies:
 - numpy
 - openmpi
 - cloudpickle =3.1.1
-- mpi4py =4.0.1
+- mpi4py =4.0.3
 - pyzmq =26.2.1
 - h5py =3.13.0
 - networkx =3.4.2
