+++ .ci_support/environment-mpich.yml
@@ -11,5 +11,5 @@ dependencies:
 - networkx =3.4.2
 - pygraphviz =1.14
 - ipython =9.0.0
-- pysqa =0.2.3
+- pysqa =0.2.4
 - setuptools =75.1.0
