### 2025-03-07 15abfd4
Bump pysqa from 0.2.3 to 0.2.4 (#600)
* Bump pysqa from 0.2.3 to 0.2.4

Bumps [pysqa](https://github.com/pyiron/pysqa) from 0.2.3 to 0.2.4.
- [Release notes](https://github.com/pyiron/pysqa/releases)
- [Commits](https://github.com/pyiron/pysqa/compare/pysqa-0.2.3...pysqa-0.2.4)

---
updated-dependencies:
- dependency-name: pysqa
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>

* [dependabot skip] Update environment

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[b
[…]
