+++ .ci_support/environment-openmpi.yml
@@ -10,6 +10,6 @@ dependencies:
 - h5py =3.13.0
 - networkx =3.4.2
 - pygraphviz =1.14
-- pysqa =0.2.3
+- pysqa =0.2.4
 - ipython =9.0.0
 - setuptools =75.1.0
