### 2025-03-07 7202ad8
Bump ipython from 9.0.0 to 9.0.1
Bumps [ipython](https://github.com/ipython/ipython) from 9.0.0 to 9.0.1.
- [Release notes](https://github.com/ipython/ipython/releases)
- [Commits](https://github.com/ipython/ipython/compare/9.0.0...9.0.1)

---
updated-dependencies:
- dependency-name: ipython
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
