+++ executorlib/standalone/hdf.py
@@ -104,8 +104,9 @@ def get_queue_id(file_name: Optional[str]) -> Optional[int]:
 
 def get_cache_data(cache_directory: str) -> list[dict]:
     file_lst = []
-    for file_name in os.listdir(cache_directory):
-        with h5py.File(os.path.join(cache_directory, file_name), "r") as hdf:
+    for task_key in os.listdir(cache_directory):
+        file_name = os.path.join(cache_directory, task_key, "cache.h5in")
+        with h5py.File(file_name, "r") as hdf:
             file_content_dict = {
                 key: cloudpickle.loads(np.void(hdf["/" + key]))
                 for key in group_dict.values()
