+++ executorlib/interactive/shared.py
@@ -148,9 +148,9 @@ def _execute_task_with_cache(
         fn_kwargs=task_dict["kwargs"],
         resource_dict=task_dict.get("resource_dict", {}),
     )
-    os.makedirs(cache_directory, exist_ok=True)
-    file_name = os.path.join(cache_directory, task_key + ".h5out")
-    if task_key + ".h5out" not in os.listdir(cache_directory):
+    os.makedirs(os.path.join(cache_directory, task_key), exist_ok=True)
+    file_name = os.path.join(cache_directory, task_key, "cache.h5out")
+    if not (task_key in os.listdir(cache_directory) and "cache.h5out" in os.listdir(os.path.join(cache_directory, task_key))):
         f = task_dict.pop("future")
         if f.set_running_or_notify_cancel():
             try:
