+++ executorlib/interactive/shared.py
@@ -150,7 +150,10 @@ def _execute_task_with_cache(
     )
     os.makedirs(os.path.join(cache_directory, task_key), exist_ok=True)
     file_name = os.path.join(cache_directory, task_key, "cache.h5out")
-    if not (task_key in os.listdir(cache_directory) and "cache.h5out" in os.listdir(os.path.join(cache_directory, task_key))):
+    if not (
+        task_key in os.listdir(cache_directory)
+        and "cache.h5out" in os.listdir(os.path.join(cache_directory, task_key))
+    ):
         f = task_dict.pop("future")
         if f.set_running_or_notify_cancel():
             try:
