### 2025-03-09 7ae04ec
set job name based on file name
