+++ executorlib/cache/queue_spawner.py
@@ -68,7 +68,7 @@ def execute_with_pysqa(
             if k in resource_dict:
                 del resource_dict[k]
         if "job_name" not in resource_dict:
-            resource_dict["job_name"] = "pysqa"
+            resource_dict["job_name"] = os.path.basename(os.path.dirname(os.path.abspath(cwd)))
         submit_kwargs.update(resource_dict)
         queue_id = qa.submit_job(**submit_kwargs)
         dump(file_name=file_name, data_dict={"queue_id": queue_id})
