### 2025-03-09 3d95236
Bump ipython from 9.0.0 to 9.0.2 (#602)
* Bump ipython from 9.0.0 to 9.0.2

Bumps [ipython](https://github.com/ipython/ipython) from 9.0.0 to 9.0.2.
- [Release notes](https://github.com/ipython/ipython/releases)
- [Commits](https://github.com/ipython/ipython/compare/9.0.0...9.0.2)

---
updated-dependencies:
- dependency-name: ipython
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>

* [dependabot skip] Update environment

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependa
[…]
