+++ .ci_support/environment-openmpi.yml
@@ -11,5 +11,5 @@ dependencies:
 - networkx =3.4.2
 - pygraphviz =1.14
 - pysqa =0.2.4
-- ipython =9.0.0
+- ipython =9.0.2
 - setuptools =75.1.0
