+++ .ci_support/environment-docs.yml
@@ -9,7 +9,7 @@ dependencies:
 - openmpi
 - cloudpickle =3.1.1
 - mpi4py =4.0.1
-- pyzmq =26.2.1
+- pyzmq =26.3.0
 - flux-core
 - jupyter-book =1.0.0
 - h5py =3.13.0
