### 2025-03-12 c8fcba4
Bump pyzmq from 26.2.1 to 26.3.0 (#604)
* Bump pyzmq from 26.2.1 to 26.3.0

Bumps [pyzmq](https://github.com/zeromq/pyzmq) from 26.2.1 to 26.3.0.
- [Release notes](https://github.com/zeromq/pyzmq/releases)
- [Commits](https://github.com/zeromq/pyzmq/compare/v26.2.1...v26.3.0)

---
updated-dependencies:
- dependency-name: pyzmq
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>

* [dependabot skip] Update environment

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot
[…]
