+++ .ci_support/environment-mpich.yml
@@ -6,7 +6,7 @@ dependencies:
 - mpich
 - cloudpickle =3.1.1
 - mpi4py =4.0.1
-- pyzmq =26.2.1
+- pyzmq =26.3.0
 - h5py =3.13.0
 - networkx =3.4.2
 - pygraphviz =1.14
