+++ pyproject.toml
@@ -26,7 +26,7 @@ classifiers = [
 ]
 dependencies = [
     "cloudpickle==3.1.1",
-    "pyzmq==26.2.1",
+    "pyzmq==26.3.0",
 ]
 dynamic = ["version"]
 
