+++ .ci_support/environment-win.yml
@@ -5,7 +5,7 @@ dependencies:
 - numpy
 - msmpi
 - cloudpickle =3.1.1
-- mpi4py =4.0.1
+- mpi4py =4.0.3
 - pyzmq =26.3.0
 - h5py =3.13.0
 - networkx =3.4.2
