### 2025-03-16 #7782 @lwshanbd
> 👋 [@lwshanbd](https://github.com/lwshanbd) -ç? What's blocking your progress?

Hi, I have installed the single-node version and it works well. But it failed on a cluster version. My initial guess is that there is something wrong with the cluster setup I tested on, and I will test the speedup on another cluster very soon.
