### 2025-03-17 #7782 @IBCHgenomic
> 👋 [@IBCHgenomic](https://github.com/IBCHgenomic) - Did you change your username from @sciencegenome? If so, I need to change this in the JOSS system to ensure notifications work right and you get credit. Also, this will enable you to generate a checklist and then to do your review.

@danielskatz yes i moved to a new job offer in Poznan, Poland and I will do the same. I hope to complete this during this week. Thank you all, Gaurav
