### 2025-03-24 f06b989
Corrections for JOSS submission
