### 2025-03-24 #7782 @jan-janssen
> 👋 [@svchb](https://github.com/svchb) - I see you created [an issue with some concerns](https://github.com/pyiron/executorlib/issues/594). [@jan-janssen](https://github.com/jan-janssen), have you looked at this and started working on it? And [@svchb](https://github.com/svchb) - can you make progress elsewhere in the review in the meantime?

@svchb I opened a pull request to address your remarks https://github.com/pyiron/executorlib/pull/606 . I look forward to your feedback.
