### 2025-03-24 #7782 @lwshanbd
Hi @jan-janssen, I tried the executorlib on `dane`@LC.

For single-node MPI execution, since there is no mpiexec available on dane, I wasn’t able to run the MPI code directly.

For multi-node execution, it crashed with the following error: `ValueError: numpy.dtype size changed, may indicate binary incompatibility. Expected 96 from C header, got 88 from PyObject`. Could you help me figure this out?

BTW, should I open an issue in the original repository?
