### 2025-03-24 #7782 @jan-janssen
> BTW, should I open an issue in the original repository?

Yes, I guess this would be easier. Somehow it seems to be using different numpy versions.
