### 2025-03-25 000a555
Corrections for JOSS submission (#606)
* Corrections for JOSS submission

* Update paper.md

* Update paper.md
