### 2025-03-25 #7782 @svchb
@jan-janssen

I also executed the examples locally and everything worked great. I only have these additional comments:
- A more comprehensive example of an actual use case would be great
- I was not able to find community guidelines
