### 2025-03-25 #7782 @jan-janssen
@svchb

> A more comprehensive example of an actual use case would be great

Do you mean adding a production example to the documentation? Or do you mean providing an example in the text of the publication? I originally had an example in the paper but then the text was too long. So I am going to add a more extensive example to the documentation.

> I was not able to find community guidelines

Do you mean in terms of a code of conduct? I added a code of conduct in https://github.com/pyiron/executorlib/pull/615
