### 2025-03-25 #7782 @svchb
@jan-janssen

> Do you mean adding a production example to the documentation? Or do you mean providing an example in the text of the publication? I originally had an example in the paper but then the text was too long. So I am going to add a more extensive example to the documentation.

Yes not in the paper either in the documentation or a link from the documentation to an example file.

> Do you mean in terms of a code of conduct? I added a code of conduct in https://github.com/pyiron/executorlib/pull/615

To quote from the review list:
Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
Usually people add a section(s) to the readme to address these 3 points.
