+++ .github/workflows/pipeline.yml
@@ -167,9 +167,11 @@ jobs:
       - name: Notebooks
         shell: bash -l {0}
         timeout-minutes: 5
-        run: >
-          flux start
-          .ci_support/build_notebooks.sh
+        run: |
+          papermill notebooks/1-single-node.ipynb notebooks/1-single-node-out.ipynb -k python3
+          flux start papermill notebooks/2-hpc-cluster.ipynb notebooks/2-hpc-cluster-out.ipynb -k python3
+          flux start papermill notebooks/3-hpc-job.ipynb notebooks/3-hpc-job-out.ipynb -k python3
+          papermill notebooks/4-developer.ipynb notebooks/4-developer-out.ipynb -k python3
 
   unittest_flux_mpich:
     needs: [black]
