+++ README.md
@@ -146,9 +146,11 @@ as hierarchical job scheduler within the allocations.
   * [Python Version](https://executorlib.readthedocs.io/en/latest/trouble_shooting.html#python-version)
   * [Resource Dictionary](https://executorlib.readthedocs.io/en/latest/trouble_shooting.html#resource-dictionary)
   * [SSH Connection](https://executorlib.readthedocs.io/en/latest/trouble_shooting.html#ssh-connection)
-* [Developer](https://executorlib.readthedocs.io/en/latest/4-developer.html)
-  * [Communication](https://executorlib.readthedocs.io/en/latest/4-developer.html#communication)
-  * [External Executables](https://executorlib.readthedocs.io/en/latest/4-developer.html#external-executables)
+* [Support & Contribution](https://executorlib.readthedocs.io/en/latest/4-developer.html)
+  * [Issues](https://executorlib.readthedocs.io/en/latest/4-developer.html#issues)
+  * [Pull Requests](https://executorlib.readthedocs.io/en/latest/4-developer.html#pull-requests)
   * [License](https://executorlib.readthedocs.io/en/latest/4-developer.html#license)
   * [Modules](https://executorlib.readthedocs.io/en/latest/4-developer.html#modules)
+  * [Communication](https://executorlib.readthedocs.io/en/latest/4-developer.html#communication)
+  * [External Executables](https://executorlib.readthedocs.io/en/latest/4-developer.html#external-executables)
 * [Interface](https://executorlib.readthedocs.io/en/latest/api.html)
