### 2025-03-28 405c472
Fix threads per core when submitting to queue
