### 2025-03-28 8d5ceb0
Add integration tests for application notebooks
