+++ .github/workflows/pipeline.yml
@@ -180,8 +180,6 @@ jobs:
       - uses: actions/checkout@v4
       - name: Merge Notebook environment
         run: |
-          cp binder/environment.yml environment.yml
-          tail --lines=+4 .ci_support/environment-integration.yml >> environment.yml
           echo -e "channels:\n  - conda-forge\n" > .condarc
       - name: Setup Mambaforge
         uses: conda-incubator/setup-miniconda@v3
@@ -189,7 +187,7 @@ jobs:
           python-version: "3.12"
           miniforge-version: latest
           condarc-file: .condarc
-          environment-file: environment.yml
+          environment-file: .ci_support/environment-integration.yml
       - name: Install
         shell: bash -l {0}
         run: |
