### 2025-03-28 8f1026f
Improved error handling (#620)
