+++ .github/workflows/pipeline.yml
@@ -199,7 +199,7 @@ jobs:
         run: |
           flux start papermill notebooks/5-1-flux-cluster-gpaw-bulk-modulus.ipynb notebooks/5-1-flux-cluster-gpaw-bulk-modulus-out.ipynb -k python3
           flux start papermill notebooks/5-2-flux-job-gpaw-bulk-modulus.ipynb notebooks/5-2-flux-job-gpaw-bulk-modulus-out.ipynb -k python3
-          flux start papermill notebooks/5-3-flux-cluster-qe-bulk-modulus.ipynb notebooks/5-3-flux-cluster-qe-bulk-modulus-out.ipynb -k python3
+          # flux start papermill notebooks/5-3-flux-cluster-qe-bulk-modulus.ipynb notebooks/5-3-flux-cluster-qe-bulk-modulus-out.ipynb -k python3
           flux start papermill notebooks/5-4-flux-job-qe-bulk-modulus.ipynb notebooks/5-4-flux-job-qe-bulk-modulus-out.ipynb -k python3
         env:
           ESPRESSO_PSEUDO: tests/integration
