### 2025-03-28 b8daa95
fix pseudo potential directory
