### 2025-03-28 65bf1da
cluster error handling
