+++ .github/workflows/pipeline.yml
@@ -201,11 +201,6 @@ jobs:
           flux start papermill notebooks/5-2-flux-job-gpaw-bulk-modulus.ipynb notebooks/5-2-flux-job-gpaw-bulk-modulus-out.ipynb -k python3
           flux start papermill notebooks/5-3-flux-cluster-qe-bulk-modulus.ipynb notebooks/5-3-flux-cluster-qe-bulk-modulus-out.ipynb -k python3
           flux start papermill notebooks/5-4-flux-job-qe-bulk-modulus.ipynb notebooks/5-4-flux-job-qe-bulk-modulus-out.ipynb -k python3
-        env:
-          ESPRESSO_PSEUDO: tests/integration
-          OMPI_MCA_plm: isolated
-          OMPI_MCA_btl_vader_single_copy_mechanism: none
-          OMPI_MCA_rmaps_base_oversubscribe: yes
 
   unittest_flux_mpich:
     needs: [black]
