### 2025-03-28 73dca53
extend error handling
