### 2025-03-28 cea1a4e
fix testing with flux
