### 2025-03-28 733f157
in interactive mode only successful calculation are stored
