+++ executorlib/interactive/shared.py
@@ -173,10 +173,7 @@ def _execute_task_with_cache(
     else:
         _, no_error_flag, result = get_output(file_name=file_name)
         future = task_dict["future"]
-        if no_error_flag:
-            future.set_result(result)
-        else:
-            future.set_exception(result)
+        future.set_result(result)
         _task_done(future_queue=future_queue)
 
 
