### 2025-03-28 c4ce1e1
Cluster error handling (#621)
* cluster error handling

* fix error read

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add additional tests

* extend error handling

* folder error

* fix testing with flux

* another test

* hit cache again

* in interactive mode only successful calculation are stored

* clean up

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
