+++ docs/application.md
@@ -1,2 +1,6 @@
 # Application
-Calculate the bulk modulus 
+While `executorlib` is designed to up-scale any Python function for high performance computing (HPC), it was initially
+developed to accelerate atomistic computational materials science simulation. To demonstrate the usage of `executorlib`
+in the context of atomistic simulation, it is combined with [atomistics](https://atomistics.readthedocs.io/) and the
+[atomic simulation environment (ASE)](https://wiki.fysik.dtu.dk/ase/) to calculate the bulk modulus with two density
+functional theory simulation codes [gpaw](https://gpaw.readthedocs.io/index.html) and [quantum espresso](https://www.quantum-espresso.org).
