+++ docs/_toc.yml
@@ -5,10 +5,10 @@ chapters:
 - file: 1-single-node.ipynb
 - file: 2-hpc-cluster.ipynb
 - file: 3-hpc-job.ipynb
-- file: trouble_shooting.md
 - file: application.md
   sections:
     - file: 5-1-gpaw.ipynb
     - file: 5-2-quantum-espresso.ipynb
+- file: trouble_shooting.md
 - file: 4-developer.ipynb
 - file: api.rst
