### 2025-03-28 3cbd96f
Add integration tests for application notebooks (#619)
* Add integration tests for application notebooks

* remove pmix

* remove constraints

* fix notebook names

* increase time out

* remove ASE command

* Disable qe cluster test

* fix pseudo potential directory

* relative to working directory

* try absolute path

* Update pipeline.yml

* Test new debugging capability

* use flux run to call pw.x

* try p2.x with mpi

* mpi takes too long

* enable 3rd test

* always nest flux

* no more nesting

* no more nesting

* merge notebooks

* fix pseudo potential directory

* add quantum espre
[…]
