+++ README.md
@@ -139,6 +139,9 @@ as hierarchical job scheduler within the allocations.
   * [SLURM](https://executorlib.readthedocs.io/en/latest/3-hpc-job.html#slurm)
   * [SLURM with Flux](https://executorlib.readthedocs.io/en/latest/3-hpc-job.html#slurm-with-flux)
   * [Flux](https://executorlib.readthedocs.io/en/latest/3-hpc-job.html#flux)
+* [Application](https://executorlib.readthedocs.io/en/latest/application.html)
+  * [GPAW](https://executorlib.readthedocs.io/en/latest/5-1-gpaw.html)
+  * [Quantum Espresso](https://executorlib.readthedocs.io/en/latest/5-2-quantum-espresso.html)
 * [Trouble Shooting](https://executorlib.readthedocs.io/en/latest/trouble_shooting.html)
   * [Filesystem Usage](https://executorlib.readthedocs.io/en/latest/trouble_shooting.html#filesystem-usage)
   * [Firewall Issues](https://executorlib.readthedocs.io/en/latest/trouble_shooting.html#firewall-issues)
