### 2025-03-28 #7782 @jan-janssen
@svchb I added a more comprehensive example in https://executorlib.readthedocs.io/en/latest/5-1-gpaw.html
