### 2025-03-29 #7782 @svchb
@jan-janssen Thanks interesting I did not think about this application.
