### 2025-03-29 #7782 @danielskatz
👋 @lwshanbd - I see you've completed your checklist, and your comments were addressed in https://github.com/openjournals/joss-reviews/issues/7782#issuecomment-2750529017

Are you ok for this to be published?
