### 2025-03-29 c15c1c1
Add example for get_cache_data()
