+++ executorlib/__init__.py
@@ -8,6 +8,7 @@ from executorlib.interfaces.slurm import (
     SlurmClusterExecutor,
     SlurmJobExecutor,
 )
+from executorlib.standalone.hdf import get_cache_data
 
 __version__ = _get_versions()["version"]
 __all__: list = [
@@ -16,4 +17,5 @@ __all__: list = [
     "SingleNodeExecutor",
     "SlurmJobExecutor",
     "SlurmClusterExecutor",
+    "get_cache_data",
 ]
