+++ executorlib/__init__.py
@@ -8,7 +8,13 @@ from executorlib.interfaces.slurm import (
     SlurmClusterExecutor,
     SlurmJobExecutor,
 )
-from executorlib.standalone.hdf import get_cache_data
+
+try:
+    from executorlib.standalone.hdf import get_cache_data
+except ImportError:
+    hdf_lst: list = []
+else:
+    hdf_lst: list = [get_cache_data]
 
 __version__ = _get_versions()["version"]
 __all__: list = [
@@ -17,5 +23,4 @@ __all__: list = [
     "SingleNodeExecutor",
     "SlurmJobExecutor",
     "SlurmClusterExecutor",
-    "get_cache_data",
-]
+] + hdf_lst
