### 2025-03-29 55286eb
Add JOSS and github stars badge to readme (#623)
