### 2025-03-29 85f33bc
extend type hint
