### 2025-03-29 5de07ac
Add example for get_cache_data() (#625)
* Add example for get_cache_data()

* fix minimal

* fix list of tools

* hid list

* add list step by step

* add to list

* ignore

* add strings

* extend type hint

* fix test to increase coverage
