### 2025-04-01 #7782 @danielskatz
@jan-janssen - I've suggested a number of small changes in https://github.com/pyiron/executorlib/pull/627 - please merge this, or let me know what you disagree with.

Additionally, JOSS authors typically use first names, not just initials. Feel free to change this in your paper if you want.
