### 2025-04-01 9bbf680
minor changes for JOSS publication (#627)
* minor changes in paper.md

* minor changes in paper.bib
