+++ docs/paper/paper.bib
@@ -20,7 +20,7 @@
 
 @inproceedings{slurm,
     author = {Morris A. Jette and Andy B. Yoo and Mark Grondona},
-    title = {SLURM: Simple Linux Utility for Resource Management},
+    title = {SLURM: Simple {L}inux Utility for Resource Management},
     booktitle = {In Lecture Notes in Computer Science: Proceedings of Job Scheduling Strategies for Parallel Processing ({JSSPP}) 2003},
     year = {2002},
     pages = {44--60},
@@ -88,7 +88,7 @@
 @inproceedings{dask,
     author = {Matthew Rocklin},
     title = {Dask: Parallel Computation with Blocked algorithms and Task Scheduling},
-    booktitle = {Proceedings of the 14th Python in Science Conference},
+    booktitle = {Proceedings of the 14th {P}ython in Science Conference},
     pages = {126 - 132},
     year = {2015},
     editor = {Kathryn Huff and James Bergstra},
@@ -97,7 +97,7 @@
 
 @inproceedings{parsl, 
     author = {Babuji, Yadu and Woodard, Anna and Li, Zhuozhao and Katz, Daniel S. and Clifford, Ben and Kumar, Rohan and Lacinski, Lukasz and Chard, Ryan and Wozniak, Justin M. and Foster, Ian and Wilde, Michael and Chard, Kyle}, 
-    title = {Parsl: Pervasive Parallel Programming in Python}, 
+    title = {Parsl: Pervasive Parallel Programming in {P}ython}, 
     year = {2019}, 
     isbn = {9781450366700}, 
     publisher = {Association for Computing Machinery}, 
@@ -136,7 +136,7 @@
 }
 
 @article{mpi4py,
-    title = {MPI for Python},
+    title = {MPI for {P}ython},
     journal = {Journal of Parallel and Distributed Computing},
     volume = {65},
     number = {9},
