### 2025-04-01 #7782 @jan-janssen
Hi @danielskatz,

I released version 1.0.0 https://github.com/pyiron/executorlib/releases/tag/executorlib-1.0.0 and it is also archived on Zenodo https://doi.org/10.5281/zenodo.15121422 .

Do you need anything else from my side?

Best,

Jan
