### 2025-04-01 #7782 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
