### 2025-04-01 e811765
Create CITATION.cff (#626)
* Create CITATION.cff

* Update CITATION.cff
